Use '+' to concat so that a new string is created
If we use concat, it modifies the original string and that change
carries over to the Product class title which would now have the
variants name concatenated to the original product title name. This is
not desired hence using ‘+’ for concatenation. Earlier concat was used
with the intention that it’s more performant than ‘+’ as it doesn’t
create a new string for every concatenation.