Refactor code to get full title related details of each product from cart
Getting the full title details of products is the responsibility of the
cart as they are compose of product variants. The store doesn’t need to
know that variants have an attribute called full_title. Also if
anything changes wrt Variants in the future, one needn’t have to make
changes in cart.rb, store.rb and variant.rb. They’d only have to make
changes in cart.rb and variant.rb.