Refactor code to use the appropriate method name to get total weight
If we use the method name total_weight_in_grams, it sounds more
specific and there isn’t a need for us to be that specific. Also, if we
want to be that specific, even the total cost that we calculate wrt
variants via the total_cost method should be renamed to
total_cost_in_dollars.

What Variant is currently doing is getting the value wrt it’s inherent
attributes. If we say actually converted the inherent unit of weight
that is currently used by variant(i.e, the weight in grams), to
something like kilograms, it would make sense to rename the method to
total_weight_in_kilograms but since that’s not the case, we needn’t be
that explicit with the method name to get total weight.